fix :pushpin: name refactor
클래스 이름 수정
