+++ BE/src/main/java/com/codesquad/baseball07/dao/PitchDao.java
@@ -1,6 +1,5 @@
 package com.codesquad.baseball07.dao;
 
-import com.codesquad.baseball07.dto.PitchDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -8,12 +7,12 @@ import org.springframework.stereotype.Repository;
 import javax.sql.DataSource;
 
 @Repository
-public class PitchDAO {
+public class PitchDao {
 
     private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     @Autowired
-    public PitchDAO(DataSource dataSource) {
+    public PitchDao(DataSource dataSource) {
         this.namedParameterJdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
     }
